Solution to date incorrect format
I finally understood how the str_to_date() function worked. The 2nd parameter is just meant to represent the format in which the string is being input. I had interpreted it as the format which we wanted to be output. I have also learned that the date_format() function is what is used to manipulate the actual formatting.